requirements.txt
# Core LLM and Agent Framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.2.0
langchain-text-splitters>=0.3.0
anthropic>=0.34.0

# Vector Database and Embeddings
chromadb>=0.5.0
sentence-transformers>=3.0.0

# PDF Processing
pypdf>=4.0.0
pdfplumber>=0.11.0
pymupdf>=1.24.0

# Structured Database
sqlalchemy>=2.0.0
neo4j>=5.0.0  # For knowledge graph option

# UI
gradio>=4.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
ruff>=0.5.0
